+++ lib/client/widgets/Login.widget.dart
@@ -36,6 +36,7 @@ class LoginWidget extends StatefulWidget {
 class _LoginWidgetState extends State<LoginWidget> {
   @override
   void initState() {
+    inputState.reset();
     super.initState();
   }