+++ lib/client/widgets/Register.widget.dart
@@ -53,6 +53,8 @@ class _RegisterWidgetState extends State<RegisterWidget> {
   @override
   void initState() {
     super.initState();
+
+    inputState.reset();
   }
 
   @override