+++ CHANGELOG.md
@@ -6,4 +6,6 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [1.0.2]
 
-- []()
+- remove expanded
+- prevent double botton sheet
+- reset login